Move repro to `dashboard-filters`
