+++ frontend/test/metabase/scenarios/question/nested.cy.spec.js
@@ -7,7 +7,6 @@ import {
   visualize,
   getDimensionByName,
   summarize,
-  visitDashboard,
   startNewQuestion,
 } from "__support__/e2e/cypress";
 
@@ -188,52 +187,6 @@ describe("scenarios > question > nested", () => {
       .contains("Average of Total")
       .closest(".List-item")
       .contains("Auto binned");
-  });
-
-  it("should show all filter options for a nested question (metabase#13186)", () => {
-    const nativeQuestionDetails = {
-      name: "13816_Q1",
-      native: {
-        query: "SELECT * FROM PRODUCTS",
-      },
-    };
-
-    cy.createNativeQuestion(nativeQuestionDetails).then(
-      ({ body: { id: Q1_ID } }) => {
-        cy.log("Convert Q1 to `query` and save as Q2");
-        cy.createQuestion({
-          name: "13816_Q2",
-          query: {
-            "source-table": `card__${Q1_ID}`,
-          },
-        });
-      },
-    );
-
-    cy.createDashboard().then(({ body: { id: DASHBOARD_ID } }) => {
-      visitDashboard(DASHBOARD_ID);
-    });
-
-    // Add Q2 to that dashboard
-    cy.icon("pencil").click();
-    cy.icon("add")
-      .last()
-      .click();
-    cy.findByText("13816_Q2").click();
-
-    // Add filter to the dashboard...
-    cy.icon("filter").click();
-    cy.findByText("Text or Category").click();
-    cy.findByText("Dropdown").click();
-    // ...and try to connect it to the question
-    cy.findByText("Select…").click();
-
-    cy.log("Reported failing in v0.36.4 (`Category` is missing)");
-    popover().within(() => {
-      cy.findByText(/Category/i);
-      cy.findByText(/Title/i);
-      cy.findByText(/Vendor/i);
-    });
   });
 
   it("should apply metrics including filter to the nested question (metabase#12507)", () => {
